+++ tests/marshal/marshal_primitive_test.py
@@ -1,5 +1,6 @@
 import mock
 import pytest
+from bravado_core.exception import SwaggerMappingError
 
 from bravado_core.marshal import marshal_primitive
 
@@ -44,6 +45,6 @@ def test_required_failure():
         'type': 'integer',
         'required': True,
     }
-    with pytest.raises(TypeError) as excinfo:
+    with pytest.raises(SwaggerMappingError) as excinfo:
         marshal_primitive(integer_spec, None)
     assert 'is a required value' in str(excinfo.value)