+++ CHANGELOG.rst
@@ -1,4 +1,15 @@
-0.1.0 (2015-XX-XX)
+1.0.0-rc2 (2015-06-01)
 ------------------
+- Fixed file uploads when marshaling a request
+- Renamed ResponseLike to IncomingResponse
+- Fixed repr of a model when it has an attr with a unicode value
 
+1.0.0-rc1 (2015-05-26)
+------------------
+- Use basePath when matching an operation to a request
+- Refactored exception hierarchy
+- Added use_models config option
+
+0.1.0 (2015-05-13)
+------------------
 - Initial release